refactoring
1.usecase 내 메서드명 변경
2.OrderId new 연산자로 객체 생성 제한
3.주문 생성 시, 상품/재고 체크 순서 변경
4.전체 타입 불일치 해결
5.엔티티 생성자 access level protected로 설정
6.usecase의 공통 주문조회 로직을 Order Loader로 분리
7.OrderItemId 생성 방식 변경, DB에 OrderId와 함께 복합키 지정

+++ service/purchasing/src/main/java/org/icd4/commerce/domain/order/OrderItem.java
@@ -16,9 +16,9 @@ public class OrderItem {
     private final OrderId orderId;
     private final ProductId productId;
     private final String productName;
-    private final long unitPrice;
-    private final long quantity;
-    private final long itemAmount; // = unitPrice * quantity
+    private final Long unitPrice;
+    private final int quantity;
+    private final Long itemAmount; // = unitPrice * quantity
     private final Map<String, String> productOptions; // 옵션 타입 ID → 옵션 값 ID
 
     /**
@@ -29,8 +29,8 @@ public class OrderItem {
             OrderId orderId,
             ProductId productId,
             String productName,
-            long unitPrice,
-            long quantity,
+            Long unitPrice,
+            int quantity,
             Map<String, String> productOptions
     ) {
         this.orderItemId = Objects.requireNonNull(orderItemId, "orderItemId is required");
